Display words in ascending alphabetical order
Write a program that reads words from a text file and displays all the words (duplicates allowed) in ascending order alphabetical order. The words must start with a letter. The text file is passed as a command-line argument.

Note: My program does not pass the file as a command-line argument.